double -> float

+++ src/Pint.c
@@ -4,19 +4,19 @@
 
 struct effect {
   int i, j;
-  double strength;
+  float strength;
 };
 
 SEXP lasso_(SEXP X_, SEXP Y_, SEXP lambda_min_, SEXP lambda_max_,
             SEXP frac_overlap_allowed_, SEXP halt_error_diff_,
             SEXP max_interaction_distance_, SEXP use_adaptive_calibration_,
             SEXP max_nz_beta_) {
-  double *x = REAL(X_);
-  double *y = REAL(Y_);
+  float *x = REAL(X_);
+  float *y = REAL(Y_);
   SEXP dim = getAttrib(X_, R_DimSymbol);
   int n = INTEGER(dim)[0];
   int p = INTEGER(dim)[1];
-  double frac_overlap_allowed = asReal(frac_overlap_allowed_);
+  float frac_overlap_allowed = asReal(frac_overlap_allowed_);
   // int p_int = p*(p+1)/2;
   int max_interaction_distance = asInteger(max_interaction_distance_);
   int p_int = get_p_int(p, max_interaction_distance);
@@ -25,7 +25,7 @@ SEXP lasso_(SEXP X_, SEXP Y_, SEXP lambda_min_, SEXP lambda_max_,
 
   int use_adaptive_calibration = asLogical(use_adaptive_calibration_);
 
-  double halt_error_diff = asReal(halt_error_diff_);
+  float halt_error_diff = asReal(halt_error_diff_);
 
   int **X = malloc(p * sizeof(int *));
   for (int i = 0; i < p; i++)
@@ -36,9 +36,9 @@ SEXP lasso_(SEXP X_, SEXP Y_, SEXP lambda_min_, SEXP lambda_max_,
       X[i][j] = (int)(x[j + i * n]);
     }
   }
-  double *Y = malloc(n * sizeof(double));
+  float *Y = malloc(n * sizeof(float));
   for (int i = 0; i < n; i++) {
-    Y[i] = (double)y[i];
+    Y[i] = (float)y[i];
   }
 
   XMatrix xmatrix;
@@ -49,7 +49,7 @@ SEXP lasso_(SEXP X_, SEXP Y_, SEXP lambda_min_, SEXP lambda_max_,
 
   Rprintf("limiting interaction distance to %d\n", max_interaction_distance);
 
-  double *beta = simple_coordinate_descent_lasso(
+  float *beta = simple_coordinate_descent_lasso(
       xmatrix, Y, n, p, max_interaction_distance, asReal(lambda_min_),
       asReal(lambda_max_), 100, 0, frac_overlap_allowed, halt_error_diff,
       log_level, NULL, 0, use_adaptive_calibration, max_nz_beta);
